moved getBluServiceName() to ShellyThingCreator
