+++ bundles/org.openhab.binding.shelly/src/main/java/org/openhab/binding/shelly/internal/ShellyDevices.java
@@ -13,7 +13,6 @@
 package org.openhab.binding.shelly.internal;
 
 import static org.openhab.binding.shelly.internal.ShellyBindingConstants.BINDING_ID;
-import static org.openhab.binding.shelly.internal.util.ShellyUtils.substringBefore;
 
 import java.util.HashSet;
 import java.util.Map;
@@ -647,26 +646,4 @@ public class ShellyDevices {
             Map.entry(THING_TYPE_SHELLYBLUBUTTON, 1), //
             Map.entry(THING_TYPE_SHELLYBLUHT, 1), //
             Map.entry(THING_TYPE_SHELLYBLUDW, 1));
-
-    /**
-     * Generates a service name based on the provided model name and MAC address.
-     * Delimiters will be stripped from the returned MAC address.
-     *
-     * @param name Model name such as SBBT-02C or just SBDW
-     * @param mac MAC address with or without colon delimiters
-     * @return service name in the form <code>&lt;service name&gt;-&lt;mac&gt;</code>
-     */
-    public static String getBluServiceName(String model, String mac) throws IllegalArgumentException {
-        String bluClass = model.contains("-") ? substringBefore(model, "-") : model;
-        ThingTypeUID uid = THING_TYPE_BY_DEVICE_TYPE.containsKey(model) ? THING_TYPE_BY_DEVICE_TYPE.get(model)
-                : THING_TYPE_BY_DEVICE_TYPE.get(bluClass);
-        if (uid != null) {
-            String serviceName = uid.getId();
-            if (!serviceName.isEmpty()) {
-                return serviceName + "-" + mac.replaceAll(":", "").toLowerCase();
-            }
-        }
-
-        throw new IllegalArgumentException("Unsupported BLU device model " + model);
-    }
 }
